+++ src/main/java/com/filter/Statistic/Types/StatisticFloat.java
@@ -1,29 +1,29 @@
 package com.filter.Statistic.Types;
 
-public class StatisticFloat implements IStatisticTypes<Float>{
-    float mean = 0;
+public class StatisticFloat implements IStatisticTypes<Double>{
+    double mean = 0;
     int count = 0;
-    float sum = 0;
-    float max = -Float.MAX_VALUE;
-    float min = Float.MAX_VALUE;
+    Double sum = (double) 0;
+    Double max = -Double.MAX_VALUE;
+    Double min = Double.MAX_VALUE;
     @Override
     public int getCount(){return count;}
     @Override
-    public Float getSum(){return sum;}
+    public Double getSum(){return sum;}
     @Override
-    public Float getMax(){return max;}
+    public Double getMax(){return max;}
     @Override
-    public Float getMin(){return min;}
+    public Double getMin(){return min;}
     @Override
-    public float getMean() {return mean;}
+    public Double getMean() {return mean;}
     @Override
     public void setCount(Integer value){count = value;}
     @Override
-    public void setMax(Float value){max = value;}
+    public void setMax(Double value){max = value;}
     @Override
-    public void setMin(Float value){min = value;}
+    public void setMin(Double value){min = value;}
     @Override
-    public void setMean(Float value, Integer value2){mean = value / value2;}
+    public void setMean(Double value, Integer value2){mean = value / value2;}
     @Override
-    public void setSum(Float value){sum += value;}
+    public void setSum(Double value){sum += value;}
 }
